Added UserContextService to avoid dependencies

+++ NoticeBoardAPI/Services/AdvertService.cs
@@ -1,12 +1,14 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using NoticeBoardAPI.Authorization;
 using NoticeBoardAPI.Entities;
 using NoticeBoardAPI.Exceptions;
 using NoticeBoardAPI.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -16,7 +18,7 @@ namespace NoticeBoardAPI.Services
     {
         AdvertDto GetById(int id);
         IEnumerable<AdvertDto> GetAll();
-        int Create(CreateAdvertDto dto, int userId);
+        int Create(CreateAdvertDto dto);
         void Delete(int id);
         void Update(UpdateAdvertDto dto, int id);
     }
@@ -26,12 +28,15 @@ namespace NoticeBoardAPI.Services
         private readonly NoticeBoardDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _contextService;
 
-        public AdvertService(NoticeBoardDbContext dbContext, IMapper mapper, IAuthorizationService authorizationService)
+        public AdvertService(NoticeBoardDbContext dbContext, IMapper mapper,
+            IAuthorizationService authorizationService, IUserContextService contextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _authorizationService = authorizationService;
+            _contextService = contextService;
         }
         public IEnumerable<AdvertDto> GetAll()
         {
@@ -63,14 +68,14 @@ namespace NoticeBoardAPI.Services
             return advertDto; 
         }
 
-        public int Create(CreateAdvertDto dto, int userId)
+        public int Create(CreateAdvertDto dto)
         {
             var result = _dbContext.Categories.FirstOrDefault(a => a.Name.ToLower() == dto.CategoryName.ToLower());
             var advert = new Advert()
             {
                 CategoryId = result.Id,
                 Description = dto.Description,
-                UserId = userId
+                UserId = _contextService.GetUseId
             };
             _dbContext.Adverts.Add(advert);
             _dbContext.SaveChanges();
@@ -84,19 +89,33 @@ namespace NoticeBoardAPI.Services
             {
                 throw new NotFoundException("Advert doesnt exist");
             }
+            var authorizationResult = _authorizationService.AuthorizeAsync(_contextService.User, advert,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbiddenException("You dont have access");
+            }
 
             _dbContext.Adverts.Remove(advert);
             _dbContext.SaveChanges();
         }
         public void Update(UpdateAdvertDto dto, int id)
         {
-            _authorizationService.AuthorizeAsync();
 
             var advert = _dbContext.Adverts.FirstOrDefault(a => a.Id == id);
             if(advert is null)
             {
                 throw new NotFoundException("Advert doesnt exist");
             }
+            var authorizationResult = _authorizationService.AuthorizeAsync(_contextService.User, advert,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbiddenException("You dont have access");
+            }
+
             advert.Description = dto.Description;
 
             _dbContext.SaveChanges();
